% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batest.R
\name{batest}
\alias{batest}
\alias{print.batest}
\alias{plot.batest}
\title{Bland-Altman Test of Method Agreement}
\usage{
batest(formula, data, id = NULL, meth = NULL, subset, na.action, weights, ...)

\method{print}{batest}(x, digits = 2, type = c("test", "model", "both"), ...)

\method{plot}{batest}(x, ...)
}
\arguments{
\item{formula}{symbolic description of the model used to perform the Bland-Altman test of type \code{y ~ x}.
The left-hand side should specify the measurements (\code{y}) to assess the agreement.
The right-hand side should specify a factor with two levels indicating two independent groups or samples to be compared. Alternatively, multilevel factors or continuously scaled variables can be specified to perform a Bland-Altman test of association, followed by binary splitting into two subgroups.}

\item{data, subset, na.action}{arguments controlling the formula processing
via \code{\link[stats]{model.frame}}. \code{data} must be provided in long format,
i.e. with two rows per subject (or item), one for each of the measurements made with the compared methods.}

\item{weights}{optional numeric vector of weights (case/frequency weights, by default).}

\item{...}{further control arguments, passed to \code{\link[partykit]{ctree_control}}}

\item{x}{an object as returned by \code{\link[coat]{batest}}.}

\item{digits}{a numeric specifying the number of digits to display.}

\item{type}{character string specifying whether \code{"test"} statistics (default), the \code{"model"} or \code{"both"} should be printed.}
}
\value{
Object of class \code{batest} with elements
\item{\code{test}}{result of the Bland-Altman test.}
\item{\code{model}}{tree model used to perform the Bland-Altman test.}
}
\description{
Function to perform a Bland-Altman test of differences in method agreement
with single measurements per subject (or item). Additional functions are
given for printing and plotting. The function \code{\link[coat]{coat}} must
be used in case of replicate measurements, setting the arguments
\code{alpha = 1} and \code{maxdepth = 1L}.
}
\details{
The minimum number of subjects (or items) in each of the two groups
or samples that are passed to \code{batest} or defined by it must be three.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(batest)}: function to print the result of the Bland-Altman test.

\item \code{plot(batest)}: function to plot the result of the Bland-Altman test.

}}
\examples{
\dontshow{ if(!requireNamespace("MethComp")) {
  if(interactive() || is.na(Sys.getenv("_R_CHECK_PACKAGE_NAME_", NA))) {
    stop("the MethComp package is required for this example but is not installed")
  } else q() }
}
## package and data
library("coat")
data("scint", package = "MethComp")
scint$DMSA <- factor(scint$meth == "DMSA")

## two-sample BA-test
testresult <- batest(y ~ sex, data = scint, id = "item", meth = "DMSA")

## display
testresult
print(testresult, digits = 1, type = "both")
plot(testresult)

}
\references{
Karapetyan S, Zeileis A, Henriksen A, Hapfelmeier A (2025).
\dQuote{Tree models for assessing covariate-dependent method agreement with an application to physical activity measurements.}
Journal of the Royal Statistical Society Series C: Applied Statistics, Volume 74, Issue 3, June 2025, Pages 775–799.
\doi{10.1093/jrsssc/qlae077}
}
