% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batrafo.repl.pair}
\alias{batrafo.repl.pair}
\title{Transformation function for paired measurements}
\usage{
batrafo.repl.pair(data, weights, control, ...)
}
\description{
Transformation function for paired measurements
}
