% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batrafo.repl.unpair}
\alias{batrafo.repl.unpair}
\title{Transformation function for unpaired measurements}
\usage{
batrafo.repl.unpair(data, weights, control, ...)
}
\description{
Transformation function for unpaired measurements
}
