% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coat.reshape}
\alias{coat.reshape}
\title{Reshape function used to prepare the data.}
\usage{
coat.reshape(
  formula,
  data,
  id = NULL,
  meth = NULL,
  replicates = FALSE,
  paired = FALSE
)
}
\description{
Reshape function used to prepare the data.
}
