% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diffs}
\alias{diffs}
\alias{means}
\title{Convenience Functions for Bland-Altman Analysis}
\usage{
diffs(y1, y2)

means(y1, y2)
}
\arguments{
\item{y1, y2}{numeric. Vectors of numeric measurements of the same length.}
}
\value{
Numeric vector with the differences or means of \code{y1} and \code{y2},
respectively.
}
\description{
Auxiliary functions for obtain the differences and means of a measurement
pair, as used in the classic Bland-Altman analysis.
}
\examples{
## pair of measurements
y1 <- 1:4
y2 <- c(2, 2, 1, 3)

## differences and means
diffs(y1, y2)
means(y1, y2)
}
