% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAtest.R
\name{BAtest}
\alias{BAtest}
\alias{print.BAtest}
\alias{plot.BAtest}
\title{Two-Sample Bland-Altman Test of Method Agreement}
\usage{
BAtest(y1, y2, groups, ...)

\method{print}{BAtest}(x, digits = 2, ...)

\method{plot}{BAtest}(x, statistics = TRUE, digits = 2, xlim.max = NULL, level = 0.95, ...)
}
\arguments{
\item{y1}{a numeric vector of measurements by one method.}

\item{y2}{a numeric vector of measurements by another method.}

\item{groups}{a factor with two levels indicating the two independent groups or samples to be compared.}

\item{...}{further arguments passed to \code{\link[partykit]{ctree_control}}.}

\item{x}{an object as returned by \code{\link[coat]{BAtest}}.}

\item{digits}{a numeric specifying the number of digits to display.}

\item{statistics}{a logical indicating whether plotting of test statistics is required.}

\item{xlim.max}{an optional numeric value to define the upper limit of the x-axis.}

\item{level}{a numeric specifying the desired coverage of the prediction interval.}
}
\value{
Object of class \code{BAtest} with elements
\item{\code{test}}{result of the Bland-Altman test.}
\item{\code{model}}{tree model used to perform the Bland-Altman test.}
}
\description{
Function to perform a two-sample Bland-Altman test for hypothesis testing of differences in method agreement. Additional functions are given for printing and plotting.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(BAtest)}: function to print the result of the Bland-Altman test.

\item \code{plot(BAtest)}: function to plot the result of the Bland-Altman test.

}}
\examples{
\dontshow{ if(!requireNamespace("MethComp")) {
  if(interactive() || is.na(Sys.getenv("_R_CHECK_PACKAGE_NAME_", NA))) {
    stop("the MethComp package is required for this example but is not installed")
  } else q() }
}
### load package ###
library("coat")

### data ###
data("VitCap", package = "MethComp")
## transform data to required 'wide' format
VitCap_wide <- reshape(VitCap, v.names = "y", timevar = "instrument",
                       idvar = c("item", "user"), drop = "meth", direction = "wide")

### perform a two-sample BA-test ###
testresult <- BAtest(VitCap_wide$y.St, VitCap_wide$y.Exp, VitCap_wide$user)
testresult

### plot ###
plot(testresult)
## without table of statistics and with expanded xlim to avoid overlapping labeling
plot(testresult, digits = 2, statistics = FALSE, xlim.max = 4800)

}
