% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coat.R
\name{coat}
\alias{coat}
\alias{coat.fit}
\title{Conditional Method Agreement Trees (COAT)}
\usage{
coat(
  formula,
  data,
  id = NULL,
  meth = NULL,
  subset,
  na.action,
  weights,
  means = FALSE,
  type = c("ctree", "mob"),
  replicates = FALSE,
  paired = FALSE,
  minsize = 10L,
  alpha = 0.05,
  minbucket = minsize,
  minsplit = NULL,
  ...
)

coat.fit(
  formula,
  data,
  subset,
  na.action,
  weights,
  means = FALSE,
  type = c("ctree", "mob"),
  replicates = FALSE,
  paired = FALSE,
  minsize = 10L,
  alpha = 0.05,
  minbucket = minsize,
  minsplit = NULL,
  ...
)
}
\arguments{
\item{formula}{symbolic description of the model of type \code{y ~ x1 + ... + xk}.
The left-hand side should specify the measurements (\code{y}) for the Bland-Altman analysis.
The right-hand side can specify any number of potential split variables for the tree.}

\item{data, subset, na.action}{arguments controlling the formula processing
via \code{\link[stats]{model.frame}}.}

\item{weights}{optional numeric vector of weights (case/frequency weights, by default).}

\item{means}{logical. Should the intra-individual mean values of measurements
be included as potential split variable?}

\item{type}{character string specifying the type of tree to be fit. Either \code{"ctree"} (default) or \code{"mob"}.}

\item{minsize, minbucket}{integer. The minimum number of observations in a subgroup.
Only one of the two arguments should be used (see also below).}

\item{minsplit}{integer. The minimum number of observations to consider splitting.
Must be at least twice the minimal subgroup size (\code{minsplit} or \code{minbucket}).
If set to \code{NULL} (the default) it is set to be at least 2.5 times the minimal
subgroup size.}

\item{...}{further control arguments, either passed to \code{\link[partykit]{ctree_control}}
or \code{\link[partykit]{mob_control}}, respectively.}
}
\value{
Object of class \code{coat}, inheriting either from \code{constparty} (if \code{\link[partykit]{ctree}}
is used) or \code{modelparty} (if \code{\link[partykit]{mob}} is used).
}
\description{
Tree models capturing the dependence of method agreement on covariates.
The classic Bland-Altman analysis is used for modeling method agreement
while the covariate dependency can be learned either nonparametrically
via conditional inference trees (CTree) or using model-based recursive
partitioning (MOB).
}
\details{
Conditional method agreement trees (COAT) employ unbiased
recursive partitioning in order to detect and model dependency on covariates
in the classic Bland-Altman analysis. One of two recursive partitioning techniques
can be used to find subgroups defined by splits in covariates to a pair
of measurements, either nonparametric conditional inference trees (CTree)
or parametric model-based trees (MOB). In both cases, each subgroup is associated
with two parameter estimates: the mean of the measurement difference (\dQuote{Bias})
and the corresponding sample standard deviation (\dQuote{SD}) which can be
used to construct the limits of agreement (i.e., the corresponding confidence intervals).

The minimum number of observations in a subgroup defaults to 10,
so that the mean and variance of the measurement differences can be estimated
reasonably for the Bland-Altman analysis. The default can be changed with
with the argument \code{minsize} or, equivalently, \code{minbucket}.
(The different names stem from slightly different conventions in the underlying
tree functions.) Consequently, the minimum number of observations to consider
splitting (\code{minsplit}) must be, at the very least, twice the minimum number
of observations per subgroup (which would allow only one possible split, though).
By default, \code{minsplit} is 2.5 times \code{minsize}.
Users are encouraged to consider whether for their application it is sensible
to increase or decrease these defaults. Finally, further control parameters
can be specified through the \code{...} argument, see
\code{\link[partykit]{ctree_control}} and \code{\link[partykit]{mob_control}},
respectively, for details.


To add the means of the measurement pair as a potential splitting variable,
there are also different equivalent strategies. The standard specification would
be via the \code{means} argument: \code{y ~ x1 + ..., means = TRUE}.
Alternatively, the user can also extend the formula argument via
\code{y ~ x1 + ... + means(y1, y2)}.

The SD is estimated by the usual sample standard deviation in each subgroup,
i.e., divided by the sample size \eqn{n - 1}. Note that the inference in the
MOB algorithm internally uses the maximum likelihood estimate (divided by \eqn{n})
instead so the the fluctuation tests for parameter instability can be applied.
}
\section{Functions}{
\itemize{
\item \code{coat.fit()}: fitter function for tree models.

}}
\examples{
\dontshow{ if(!requireNamespace("MethComp")) {
  if(interactive() || is.na(Sys.getenv("_R_CHECK_PACKAGE_NAME_", NA))) {
    stop("the MethComp package is required for this example but is not installed")
  } else q() }
}
### single measurements per item

## package and data
library("coat")
data("scint", package = "MethComp")

## coat based on ctree()
tr1 <- coat(y ~ age + sex, data = scint, id = "item", meth = "meth")

## coat based on mob() including mean values of paired measurements as predictor
tr2 <- coat(y ~ age + sex, data = scint, id = "item", meth = "meth", means = TRUE, type = "mob")

## display
print(tr1)
plot(tr1)

print(tr2)
plot(tr2)

## tweak various graphical arguments of the panel function (just for illustration):
## different colors, nonparametric bootstrap percentile confidence intervals, ...
plot(tr1, tp_args = list(
  xscale = c(0, 150), linecol = "deeppink",
  confint = TRUE, B = 250, cilevel = 0.5, cicol = "gold"
))
}
\references{
Karapetyan S, Zeileis A, Henriksen A, Hapfelmeier A (2025).
\dQuote{Tree models for assessing covariate-dependent method agreement with an application to physical activity measurements.}
Journal of the Royal Statistical Society Series C: Applied Statistics, Volume 74, Issue 3, June 2025, Pages 775–799.
\doi{10.1093/jrsssc/qlae077}
}
